@@ -1,4 +1,4 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -20,5 +20,6 @@ let package = Package(
             dependencies: [],
             path: "Sources"
         ),
-    ]
+    ],
+    swiftLanguageVersions: [.v6, .v5]
 )
