@@ -23,63 +23,35 @@ enum Auth {
     static let refreshTokenKey = "refreshToken"
   }
   
-  static func getAccessToken(settings: LoginSetting, completion: @escaping (Result<String, Error>) -> Void) {
-    KeychainHelper.getToken(key: Constants.accessTokenKey) { token in
-      if let token = token,
-         JWTHelper.isValid(token: token) {
-        DispatchQueue.main.async {
-          completion(.success(token))
+  static func getAccessToken(settings: LoginSetting) async throws -> String {
+    if let token = await KeychainHelper.getToken(key: Constants.accessTokenKey),
+       JWTHelper.isValid(token: token) {
+      return token
+    } else {
+      if let refreshToken = await KeychainHelper.getToken(key: Constants.accessTokenKey),
+         JWTHelper.isValid(token: refreshToken) {
+        
+        do {
+          return try await refreshAccessToken(refreshToken)
+        } catch {
+          return try await requestLogin(settings)
         }
       } else {
-        KeychainHelper.getToken(key: Constants.accessTokenKey) { refreshToken in
-          if let refreshToken = refreshToken,
-             JWTHelper.isValid(token: refreshToken) {
-            refreshAccessToken(refreshToken) { result in
-              DispatchQueue.main.async {
-                switch result {
-                case .success(let accessToken):
-                  completion(.success(accessToken))
-                case .failure(let error):
-                  requestLogin(settings, completion)
-                }
-              }
-            }
-          } else {
-            DispatchQueue.main.async {
-              requestLogin(settings, completion)
-            }
-          }
-        }
+        return try await requestLogin(settings)
       }
     }
   }
   
-  private static func requestLogin(_ settings: LoginSetting, _ completion: @escaping (Result<String, Error>) -> Void) {
-    login(settings: settings) { result in
-      switch result {
-      case .success(let response):
-        KeychainHelper.setToken(response.accessToken, key: Constants.accessTokenKey) { error in
-          if let error = error {
-            completion(.failure(error))
-            return
-          }
-          
-          KeychainHelper.setToken(response.refreshToken, key: Constants.refreshTokenKey) { error in
-            if let error = error {
-              completion(.failure(error))
-              return
-            }
-            
-            completion(.success(response.accessToken))
-          }
-        }
-      case .failure(let error):
-        completion(.failure(error))
-      }
-    }
+  private static func requestLogin(_ settings: LoginSetting) async throws -> String {
+    let result = try await login(settings: settings)
+    
+    try await KeychainHelper.setToken(result.accessToken, key: Constants.accessTokenKey)
+    try await KeychainHelper.setToken(result.refreshToken, key: Constants.refreshTokenKey)
+    
+    return result.accessToken
   }
   
-  private static func refreshAccessToken(_ refreshToken: String, completion: @escaping (Result<String, Error>) -> Void) {
+  private static func refreshAccessToken(_ refreshToken: String) async throws -> String {
     let url = URL(string: "oauth/token", relativeTo: Constants.url)!
 
     let parameters = [
@@ -93,78 +65,77 @@ enum Auth {
     request.setValue("application/json", forHTTPHeaderField: "Content-Type")
     request.httpBody = try! JSONSerialization.data(withJSONObject: parameters, options: [])
     
-    URLSession(configuration: URLSessionConfiguration.ephemeral).refreshAccessToken(request) { result in
-      switch result {
-      case .success(let response):
-        KeychainHelper.setToken(response.accessToken, key: Constants.accessTokenKey) { error in
-          if let error = error {
-            completion(.failure(error))
-            return
-          }
-          completion(.success(response.accessToken))
-        }
-      case .failure(let error):
-        completion(.failure(error))
-      }
-    }
+    let response = try await URLSession(configuration: URLSessionConfiguration.ephemeral).refreshAccessToken(request)
+    
+    try await KeychainHelper.setToken(response.accessToken, key: Constants.accessTokenKey)
+    
+    return response.accessToken
   }
 
-  private static func login(
-    settings: LoginSetting,
-    completion: @escaping (Result<AuthCodeResponse, Error>) -> Void)
+  private static func login(settings: LoginSetting) async throws -> AuthCodeResponse
   {
-    let verifier = getVerifier()!
-    let challenge = getChallenge(for: verifier)!
+    return try await withCheckedThrowingContinuation { continuation in
+      let verifier = getVerifier()!
+      let challenge = getChallenge(for: verifier)!
 
-    let authorize = URL(string: "authorize", relativeTo: Constants.url)!
-    var components = URLComponents(url: authorize, resolvingAgainstBaseURL: true)!
-    var items: [URLQueryItem] = []
-    var entries: [String: String] = [:]
+      let authorize = URL(string: "authorize", relativeTo: Constants.url)!
+      var components = URLComponents(url: authorize, resolvingAgainstBaseURL: true)!
+      var items: [URLQueryItem] = []
+      var entries: [String: String] = [:]
 
-    entries["scope"] = "openid profile email offline_access"
-    entries["client_id"] = Constants.clientId
-    entries["response_type"] = "code"
-    if case .connection(let string) = settings {
-      entries["connection"] = string
-    }
-    entries["redirect_uri"] = Constants.redirectUri.absoluteString
-    entries["state"] = generateDefaultState()
-    entries["audience"] = "https://auth0-jwt-authorizer"
-    entries.forEach { items.append(URLQueryItem(name: $0, value: $1)) }
-    components.queryItems = items
-    components.percentEncodedQuery = components.percentEncodedQuery?.replacingOccurrences(of: "+", with: "%2B")
+      entries["scope"] = "openid profile email offline_access"
+      entries["client_id"] = Constants.clientId
+      entries["response_type"] = "code"
+      if case .connection(let string) = settings {
+        entries["connection"] = string
+      }
+      entries["redirect_uri"] = Constants.redirectUri.absoluteString
+      entries["state"] = generateDefaultState()
+      entries["audience"] = "https://auth0-jwt-authorizer"
+      entries.forEach { items.append(URLQueryItem(name: $0, value: $1)) }
+      components.queryItems = items
+      components.percentEncodedQuery = components.percentEncodedQuery?.replacingOccurrences(of: "+", with: "%2B")
 
-    let url = components.url!
-    let session = ASWebAuthenticationSession(
-      url: url,
-      callbackURLScheme: Constants.redirectUri.scheme!) { url, error in
-        if let error {
-          completion(.failure(error))
-          return
-        }
+      let url = components.url!
+      let session = ASWebAuthenticationSession(
+        url: url,
+        callbackURLScheme: Constants.redirectUri.scheme!) { url, error in
+          if let error {
+            continuation.resume(throwing: error)
+            return
+          }
 
-        if let url {
-          let components = URLComponents(url: url, resolvingAgainstBaseURL: false)
-          let code = components!.queryItems!.first(where: { $0.name == "code"})
-          if let code {
-            self.exchangeAuthorizationCodeForTokens(authorizationCode: code.value!, verifier: verifier) { result in
-              completion(result)
+          if let url {
+            let components = URLComponents(url: url, resolvingAgainstBaseURL: false)
+            let code = components!.queryItems!.first(where: { $0.name == "code"})
+            if let code {
+              Task {
+                do {
+                  let response = try await self.exchangeAuthorizationCodeForTokens(authorizationCode: code.value!, verifier: verifier)
+                  continuation.resume(returning: response)
+                } catch {
+                  continuation.resume(throwing: error)
+                }
+              }
+            } else {
+              continuation.resume(throwing: LoginError.noCode)
             }
           } else {
-            completion(.failure(LoginError.noCode))
+            continuation.resume(throwing: LoginError.noUrl)
           }
-        } else {
-          completion(.failure(LoginError.noUrl))
         }
+      
+      DispatchQueue.main.async {
+        session.presentationContextProvider = PresentationContextProvider.shared
+        session.start()
       }
-    session.presentationContextProvider = PresentationContextProvider.shared
-    session.start()
+      
+    }
   }
 
   private static func exchangeAuthorizationCodeForTokens(
     authorizationCode: String,
-    verifier: String,
-    completion: @escaping (Result<AuthCodeResponse, Error>) -> Void)
+    verifier: String) async throws -> AuthCodeResponse
   {
     let url = URL(string: "oauth/token", relativeTo: Constants.url)!
 
@@ -181,7 +152,7 @@ enum Auth {
     request.setValue("application/json", forHTTPHeaderField: "Content-Type")
     request.httpBody = try! JSONSerialization.data(withJSONObject: parameters, options: [])
     
-    URLSession(configuration: URLSessionConfiguration.ephemeral).getAuthDataWith(request, completion: completion)
+    return try await URLSession(configuration: URLSessionConfiguration.ephemeral).getAuthDataWith(request)
   }
 
   private static func getVerifier() -> String? {
