@@ -9,7 +9,7 @@ import Foundation
 
 /// Type of authenticated access to required. The default case shows the Emerge Tools login page.
 /// A custom connection can be used to automatically redirect to an SSO page.
-public enum LoginSetting {
+public enum LoginSetting: Sendable {
   case `default`
   case connection(String)
 }
