@@ -15,63 +15,59 @@ enum RequestError: Error {
 }
 
 extension URLSession {
-  func checkForUpdate(_ request: URLRequest, completion: @escaping (Result<DistributionUpdateCheckResponse, Error>) -> Void) {
-    self.perform(request, decode: DistributionUpdateCheckResponse.self, useCamelCase: true, completion: completion) { [weak self] data, statusCode in
+  func checkForUpdate(_ request: URLRequest) async throws -> DistributionUpdateCheckResponse {
+    return try await self.perform(request, decode: DistributionUpdateCheckResponse.self, useCamelCase: true, decodeErrorData:  { [weak self] data, statusCode in
       return self?.getErrorFrom(data: data, statusCode: statusCode) ?? RequestError.badRequest("")
-    }
+    })
   }
   
-  func getAuthDataWith(_ request: URLRequest, completion: @escaping (Result<AuthCodeResponse, Error>) -> Void) {
-    self.perform(request, decode: AuthCodeResponse.self, useCamelCase: false, completion: completion) { _, _ in
+  func getAuthDataWith(_ request: URLRequest) async throws -> AuthCodeResponse {
+    return try await self.perform(request,
+                              decode: AuthCodeResponse.self,
+                              useCamelCase: false,
+                              decodeErrorData:  { data, statusCode in
       return RequestError.badRequest("")
-    }
+    })
   }
   
-  func refreshAccessToken(_ request: URLRequest, completion: @escaping (Result<AuthRefreshResponse, Error>) -> Void) {
-    self.perform(request, decode: AuthRefreshResponse.self, useCamelCase: false, completion: completion) { _, _ in
+  func refreshAccessToken(_ request: URLRequest) async throws -> AuthRefreshResponse {
+    return try await self.perform(request,
+                              decode: AuthRefreshResponse.self,
+                              useCamelCase: false,
+                              decodeErrorData:  { _, _ in
       return RequestError.badRequest("")
-    }
+    })
   }
   
-  func getReleaseInfo(_ request: URLRequest, completion: @escaping (Result<DistributionReleaseInfo, Error>) -> Void) {
-    self.perform(request, decode: DistributionReleaseInfo.self, useCamelCase: true, completion: completion) { [weak self] data, statusCode in
+  func getReleaseInfo(_ request: URLRequest) async throws -> DistributionReleaseInfo {
+    return try await self.perform(request,
+                              decode: DistributionReleaseInfo.self,
+                              useCamelCase: true,
+                              decodeErrorData:  { [weak self] data, statusCode in
       return self?.getErrorFrom(data: data, statusCode: statusCode) ?? RequestError.badRequest("")
-    }
+    })
   }
   
   private func perform<T: Decodable>(_ request: URLRequest,
                                      decode decodable: T.Type,
                                      useCamelCase: Bool = true,
-                                     completion: @escaping (Result<T, Error>) -> Void,
-                                     decodeErrorData: ((Data, Int) -> Error)?) {
-    URLSession.shared.dataTask(with: request) { (data, response, error) in
-      var result: Result<T, Error> = .failure(RequestError.unknownError)
-      defer {
-        completion(result)
-      }
-      if let error = error {
-        result = .failure(error)
-        return
-      }
-      guard let httpResponse = response as? HTTPURLResponse,
-            let data = data else {
-        completion(.failure(RequestError.invalidData))
-        return
-      }
-      guard (200...299).contains(httpResponse.statusCode) else {
-        let error = decodeErrorData?(data, httpResponse.statusCode) ?? RequestError.badRequest("Unknown error")
-        result = .failure(error)
-        return
-      }
+                                     decodeErrorData: (@Sendable (Data, Int) -> Error)?) async throws -> T {
+    let (data, response) = try await URLSession.shared.data(for: request)
+    
+    guard let httpResponse = response as? HTTPURLResponse else {
+      throw RequestError.invalidData
+    }
+    
+    guard (200...299).contains(httpResponse.statusCode) else {
+      let error = decodeErrorData?(data, httpResponse.statusCode) ?? RequestError.badRequest("Unknown error")
+      throw error
+    }
       
-      do {
-        let jsonDecoder = JSONDecoder()
-        jsonDecoder.keyDecodingStrategy = useCamelCase ? .useDefaultKeys : .convertFromSnakeCase
-        result = .success(try jsonDecoder.decode(decodable, from: data))
-      } catch {
-        result = .failure(error)
-      }
-    }.resume()
+    let jsonDecoder = JSONDecoder()
+    jsonDecoder.keyDecodingStrategy = useCamelCase ? .useDefaultKeys : .convertFromSnakeCase
+    return try jsonDecoder.decode(decodable, from: data)
+    
+    throw RequestError.unknownError
   }
   
   private func getErrorFrom(data: Data, statusCode: Int) -> RequestError {
