@@ -11,31 +11,36 @@ import Foundation
 import UIKit
 import ETDistribution
 
+@MainActor
 struct UpdateUtil {
   static func checkForUpdates() {
     ETDistribution.shared.checkForUpdate(params: CheckForUpdateParams(apiKey: Constants.apiKey)) { result in
-      guard case let .success(releaseInfo) = result else {
-        if case let .failure(error) = result {
-          print("Error checking for update: \(error)")
-        }
-        return
-      }
-      
-      guard let releaseInfo = releaseInfo else {
-        print("Already up to date")
-        return
-      }
-      
-      print("Update found: \(releaseInfo), requires login: \(releaseInfo.loginRequiredForDownload)")
-      if releaseInfo.loginRequiredForDownload {
-        // Get new release info, with login
-        ETDistribution.shared.getReleaseInfo(releaseId: releaseInfo.id) { newReleaseInfo in
-          if case let .success(newReleaseInfo) = newReleaseInfo {
-            UpdateUtil.installRelease(releaseInfo: newReleaseInfo)
+      Task {
+        guard case let .success(releaseInfo) = result else {
+          if case let .failure(error) = result {
+            print("Error checking for update: \(error)")
           }
+          return
         }
-      } else {
-        UpdateUtil.installRelease(releaseInfo: releaseInfo)
+        
+        guard let releaseInfo = releaseInfo else {
+          print("Already up to date")
+          return
+        }
+        
+        print("Update found: \(releaseInfo), requires login: \(releaseInfo.loginRequiredForDownload)")
+        if releaseInfo.loginRequiredForDownload {
+          // Get new release info, with login
+          await ETDistribution.shared.getReleaseInfo(releaseId: releaseInfo.id) { newReleaseInfo in
+            Task {
+              if case let .success(newReleaseInfo) = newReleaseInfo {
+                await UpdateUtil.installRelease(releaseInfo: newReleaseInfo)
+              }
+            }
+          }
+        } else {
+          await UpdateUtil.installRelease(releaseInfo: releaseInfo)
+        }
       }
     }
   }
@@ -43,52 +48,50 @@ struct UpdateUtil {
   static func checkForUpdatesWithLogin() {
     let params = CheckForUpdateParams(apiKey: Constants.apiKey, requiresLogin: true)
     ETDistribution.shared.checkForUpdate(params: params) { result in
-      guard case let .success(releaseInfo) = result else {
-        if case let .failure(error) = result {
-          print("Error checking for update: \(error)")
-        }
-        return
-      }
-      
-      guard let releaseInfo = releaseInfo else {
-        print("Already up to date")
-        return
-      }
-      
-      print("Update found: \(releaseInfo), requires login: \(releaseInfo.loginRequiredForDownload)")
-      if releaseInfo.loginRequiredForDownload {
-        // Get new release info, with login
-        ETDistribution.shared.getReleaseInfo(releaseId: releaseInfo.id) { newReleaseInfo in
-          if case let .success(newReleaseInfo) = newReleaseInfo {
-            UpdateUtil.installRelease(releaseInfo: newReleaseInfo)
+      Task {
+        guard case let .success(releaseInfo) = result else {
+          if case let .failure(error) = result {
+            print("Error checking for update: \(error)")
           }
+          return
         }
-      } else {
-        UpdateUtil.installRelease(releaseInfo: releaseInfo)
+        
+        guard let releaseInfo = releaseInfo else {
+          print("Already up to date")
+          return
+        }
+        
+        print("Update found: \(releaseInfo), requires login: \(releaseInfo.loginRequiredForDownload)")
+        if releaseInfo.loginRequiredForDownload {
+          // Get new release info, with login
+          await ETDistribution.shared.getReleaseInfo(releaseId: releaseInfo.id) { newReleaseInfo in
+            Task {
+              if case let .success(newReleaseInfo) = newReleaseInfo {
+                await UpdateUtil.installRelease(releaseInfo: newReleaseInfo)
+              }
+            }
+          }
+        } else {
+          await UpdateUtil.installRelease(releaseInfo: releaseInfo)
+        }
       }
     }
   }
   
   static func clearTokens() {
-    delete(key: "accessToken") {
-      delete(key: "refreshToken") {
-        print("Tokens cleared")
-      }
-    }
+    delete(key: "accessToken")
+    delete(key: "refreshToken")
+    print("Tokens cleared")
   }
   
-  private static func delete(key: String, completion: @escaping () -> Void) {
-    DispatchQueue.global().async {
-      let attributes = [
-        kSecClass: kSecClassGenericPassword,
-        kSecAttrService: "com.emerge.ETDistribution",
-        kSecAttrAccount: key,
-      ] as CFDictionary
+  private static func delete(key: String) {
+    let attributes = [
+      kSecClass: kSecClassGenericPassword,
+      kSecAttrService: "com.emerge.ETDistribution",
+      kSecAttrAccount: key,
+    ] as CFDictionary
 
-      SecItemDelete(attributes)
-      
-      completion()
-    }
+    SecItemDelete(attributes)
   }
   
   private static func installRelease(releaseInfo: DistributionReleaseInfo) {
